+++ context/app/static/js/components/searchPage/HierarchicalMenuFilter/index.js
@@ -1,0 +1,3 @@
+import { HierarchicalMenuFilter } from './HierarchicalMenuFilter';
+
+export default HierarchicalMenuFilter;
